Update LTE.py to avoid CGDCONT while radio enabled
This update ensures that CGDCONT is not set while the LTE modem is in CFUN=1. However, to allow the user to change the APN while the modem is attached or attaching, current CGDCONT is checked and if it differs from the provided APN value, we go to CFUN=4 and update CGDCONT before enabling the radio again.